% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_all.R
\name{deploy_all}
\alias{deploy_all}
\title{Vectorized and safe version of deploy_now}
\usage{
deploy_all(
  tenants,
  where = "development",
  servers = c(development = "devapp.shintolabs.net", production = "app.shintolabs.net",
    eindhoven_premium = "eindhoven.shintolabs.net")
)
}
\arguments{
\item{tenants}{vector of tenants to deploy}

\item{where}{Refers to config section (development/production/)}

\item{servers}{}
}
\description{
Also checks the db config file and prompts for missing passwords
}
