% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_now.R
\name{deploy_now}
\alias{deploy_now}
\title{Deploy Juno directly}
\usage{
deploy_now(
  tenant,
  appname,
  appid = NULL,
  db_config_file = "conf/config.yml",
  log_deployment = TRUE,
  launch_browser = TRUE,
  deploy_location = tempdir(),
  delete_after_deploy = TRUE,
  where = c("devapp.shintolabs.net", "app.shintolabs.net")
)
}
\arguments{
\item{tenant}{Tenant name (e.g. 'DEMO')}

\item{appname}{Name of app on posit connect (e.g. ')}

\item{appid}{Optional: the app ID (looks like a uuid in the url at posit connect). Needed if you
are a collaborator on an app and want to update it (the original uploader does not need the ID)}

\item{db_config_file}{Standard location of the DB config file}

\item{log_deployment}{If TRUE, attempts to log the deployment in DB}

\item{launch_browser}{If TRUE, opens a browser with the app after deployment}

\item{deploy_location}{App will be deployed from a temp directory, or set a directory here if
you want to inspect the pre-deploy files.}

\item{delete_after_deploy}{If TRUE, deletes all files after deploying.}

\item{where}{The server to deploy to (devapp or app)}
}
\description{
Does not make a 'deploy project' but deploys Juno for a tenant immediately,
from a temporary directory
}
