% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_db_config.R
\name{prepare_db_config}
\alias{prepare_db_config}
\title{Make sure database passwords are entered in the conf/config.yml file}
\usage{
prepare_db_config(
  tenants = NULL,
  where = "development",
  servers = c(development = "devapp.shintolabs.net", production = "app.shintolabs.net",
    eindhoven_premium = "eindhoven.shintolabs.net"),
  db_connections_required = list(data_bag = list(dbname = "data_bag", dbuser =
    "nlextract"), data_cbs = list(dbname = "data_cbs", dbuser = "data_cbs"), data_brk =
    list(dbname = "data_brk", dbuser = "nlextract"), shintousers = list(dbname =
    "shintousers", dbuser = "shintousers"), data_osm_nederland = list(dbname =
    "data_osm_nederland", dbuser = "datastraat"))
)
}
\arguments{
\item{tenants}{List of tenant names, or wbmr::get_tenant_choices()}

\item{where}{Server to deploy to, must correspond to a name in the 'servers' argument}

\item{servers}{List of servers; the name will be the name in the config section (this name corresponds to the R_CONFIG_ACTIVE environment variable on the posit connect server)'}

\item{db_connections_required}{List of database connections needed. The name is the entry in the config file, each has a list with 'dbname' and 'dbuser'}
}
\description{
Prompts for missing passwords, have 1Password ready to go!.
Also adds required database connections (data_bag, etc.), and prompts for API keys (TinyMCE, etc.)
}
