% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_gemeente_name}
\alias{normalize_gemeente_name}
\alias{make_app_name}
\alias{set_gemeente}
\title{Make code friendly version of gemeente name}
\usage{
normalize_gemeente_name(gemeente)

make_app_name(gemeente)

set_gemeente(gemeente)
}
\description{
Overwrites current this_version.yml!
}
