% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_gemeente_name}
\alias{normalize_gemeente_name}
\alias{make_app_name}
\alias{get_current_db_name}
\alias{set_gemeente}
\alias{get_gemeente}
\title{Make code friendly version of gemeente name}
\usage{
normalize_gemeente_name(gemeente)

make_app_name(gemeente)

get_current_db_name()

set_gemeente(gemeente)

get_gemeente()
}
\description{
Corresponds to the application name on rsconnect

Overwrites current this_version.yml!
}
