% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_tenant_name}
\alias{normalize_tenant_name}
\alias{make_app_name}
\alias{get_current_db_name}
\alias{set_tenant}
\alias{set_random_tenant}
\alias{get_tenant}
\title{Make code friendly version of tenant name}
\usage{
normalize_tenant_name(tenant)

make_app_name(tenant)

get_current_db_name()

set_tenant(tenant, path = getwd())

set_random_tenant()

get_tenant()
}
\arguments{
\item{tenant}{Tenant (e.g. 'DEMO')}

\item{path}{Path where to update the this_version.yml file}
}
\description{
Corresponds to the application name on rsconnect

Overwrites current this_version.yml!

Travel somewhere!
}
